move address index functions to separated files
