+++ src/rpc/addresstransaction.h
@@ -1,0 +1,11 @@
+#ifndef ADDRESSTRANSACTION_H
+#define ADDRESSTRANSACTION_H
+
+class UniValue;
+class JSONRPCRequest;
+
+UniValue searchrawtransactions(const JSONRPCRequest& request);
+
+UniValue getaddressbalance(const JSONRPCRequest& request);
+
+#endif // ADDRESSTRANSACTION_H
