+++ src/txdb.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <txdb.h>
+#include <addressindex.h>
 
 #include <chainparams.h>
 #include <hash.h>
@@ -150,10 +150,6 @@ size_t CCoinsViewDB::EstimateSize() const
 }
 
 CBlockTreeDB::CBlockTreeDB(size_t nCacheSize, bool fMemory, bool fWipe) : CDBWrapper(GetDataDir() / "blocks" / "index", nCacheSize, fMemory, fWipe) {
-    if (!Read('S', salt)) {
-        salt = GetRandHash();
-        Write('S', salt);
-    }
 }
 
 bool CBlockTreeDB::ReadBlockFileInfo(int nFile, CBlockFileInfo &info) {
@@ -253,60 +249,11 @@ bool CBlockTreeDB::WriteTxIndex(const std::vector<std::pair<uint256, CDiskTxPos>
     return WriteBatch(batch);
 }
 
-bool CBlockTreeDB::ReadAddrIndex(uint160 addrid, std::vector<CExtDiskTxPos> &list) {
-    boost::scoped_ptr<CDBIterator> pcursor(NewIterator());
-
-    uint64_t lookupid;
-    {
-        CHashWriter ss(SER_GETHASH, 0);
-        ss << salt;
-        ss << addrid;
-        lookupid = UintToArith256(ss.GetHash()).GetLow64();
-    }
-
-    pcursor->Seek(std::make_pair('a', lookupid));
-
-    while (pcursor->Valid()) {
-        std::pair<std::pair<char, uint64_t>, CExtDiskTxPos> key;
-        if (pcursor->GetKey(key) && key.first.first == 'a' && key.first.second == lookupid) {
-            list.push_back(key.second);
-        } else {
-            break;
-        }
-        pcursor->Next();
-    }
-    return true;
-}
-
-bool CBlockTreeDB::AddAddrIndex(const std::vector<std::pair<uint160, CExtDiskTxPos> > &list) {
-    unsigned char foo[0];
-    CDBBatch batch(*this);
-    for (std::vector<std::pair<uint160, CExtDiskTxPos> >::const_iterator it=list.begin(); it!=list.end(); it++) {
-        CHashWriter ss(SER_GETHASH, 0);
-        ss << salt;
-        ss << it->first;
-        batch.Write(std::make_pair(std::make_pair('a', UintToArith256(ss.GetHash()).GetLow64()), it->second), FLATDATA(foo));
-    }
-    return WriteBatch(batch, true);
-}
-
 bool CBlockTreeDB::EraseTxIndex(const std::vector<std::pair<uint256, CDiskTxPos> >&vect) {
     CDBBatch batch(*this);
     for (std::vector<std::pair<uint256,CDiskTxPos> >::const_iterator it=vect.begin(); it!=vect.end(); it++)
         batch.Erase(std::make_pair(DB_TXINDEX, it->first));
     return WriteBatch(batch);
-}
-
-bool CBlockTreeDB::EraseAddrIndex(const std::vector<std::pair<uint160, CExtDiskTxPos> > &list) {
-    unsigned char foo[0];
-    CDBBatch batch(*this);
-    for (std::vector<std::pair<uint160, CExtDiskTxPos> >::const_iterator it=list.begin(); it!=list.end(); it++) {
-        CHashWriter ss(SER_GETHASH, 0);
-        ss << salt;
-        ss << it->first;
-        batch.Erase(std::make_pair(std::make_pair('a', UintToArith256(ss.GetHash()).GetLow64()), it->second));
-    }
-    return WriteBatch(batch, true);
 }
 
 bool CBlockTreeDB::WriteFlag(const std::string &name, bool fValue) {
