+++ src/txdb.h
@@ -18,7 +18,7 @@
 class CBlockIndex;
 class CCoinsViewDBCursor;
 class uint256;
-struct CExtDiskTxPos;
+struct CExtDiskTxPos; 
 
 //! No need to periodic flush if at least this much space still available.
 static constexpr int MAX_BLOCK_COINSDB_USAGE = 10;
@@ -70,45 +70,6 @@ struct CDiskTxPos : public CDiskBlockPos
     }
 };
 
-struct CExtDiskTxPos : public CDiskTxPos
-{
-    unsigned int nHeight;
-
-    ADD_SERIALIZE_METHODS;
-
-    template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream& s, Operation ser_action) {
-            READWRITE(*(CDiskTxPos*)this);
-            READWRITE(VARINT(nHeight));
-    }
-
-    CExtDiskTxPos(const CDiskTxPos &pos, int nHeightIn) : CDiskTxPos(pos), nHeight(nHeightIn) {
-    }
-
-    CExtDiskTxPos() {
-        SetNull();
-    }
-
-    void SetNull() {
-        CDiskTxPos::SetNull();
-        nHeight = 0;
-    }
-
-    friend bool operator==(const CExtDiskTxPos &a, const CExtDiskTxPos &b) {
-        return (a.nHeight == b.nHeight && a.nFile == b.nFile && a.nPos == b.nPos && a.nTxOffset == b.nTxOffset);
-    }
-
-    friend bool operator!=(const CExtDiskTxPos &a, const CExtDiskTxPos &b) {
-        return !(a == b);
-    }
-
-    friend bool operator<(const CExtDiskTxPos &a, const CExtDiskTxPos &b) {
-        if (a.nHeight < b.nHeight) return true;
-        if (a.nHeight > b.nHeight) return false;
-        return ((const CDiskTxPos)a < (const CDiskTxPos)b);
-    }
-};
-
 /** CCoinsView backed by the coin database (chainstate/) */
 class CCoinsViewDB final : public CCoinsView
 {
@@ -154,8 +115,6 @@ private:
 /** Access to the block database (blocks/index/) */
 class CBlockTreeDB : public CDBWrapper
 {
-private:
-    uint256 salt;
 public:
     explicit CBlockTreeDB(size_t nCacheSize, bool fMemory = false, bool fWipe = false);
 
@@ -169,10 +128,7 @@ public:
     bool ReadReindexing(bool &fReindexing);
     bool ReadTxIndex(const uint256 &txid, CDiskTxPos &pos);
     bool WriteTxIndex(const std::vector<std::pair<uint256, CDiskTxPos> > &vect);
-    bool ReadAddrIndex(uint160 addrid, std::vector<CExtDiskTxPos> &list);
-    bool AddAddrIndex(const std::vector<std::pair<uint160, CExtDiskTxPos> > &list);
     bool EraseTxIndex(const std::vector<std::pair<uint256, CDiskTxPos> >&vect);
-    bool EraseAddrIndex(const std::vector<std::pair<uint160, CExtDiskTxPos> > &list);
     bool WriteFlag(const std::string &name, bool fValue);
     bool ReadFlag(const std::string &name, bool &fValue);
     bool LoadBlockIndexGuts(const Consensus::Params& consensusParams, std::function<CBlockIndex*(const uint256&)> insertBlockIndex);
