+++ src/validation.cpp
@@ -33,6 +33,7 @@
 #include <timedata.h>
 #include <tinyformat.h>
 #include <txdb.h>
+#include <addressindex.h>
 #include <txmempool.h>
 #include <ui_interface.h>
 #include <undo.h>
@@ -236,10 +237,6 @@ CScript COINBASE_FLAGS;
 
 const std::string strMessageMagic = "Bitcoin Signed Message:\n";
 
-void static BuildAddrIndex(const CScript &script, const CExtDiskTxPos &pos, std::vector<std::pair<uint160, CExtDiskTxPos> > &out);
-
-static bool EraseTxIndexDataForBlock(std::vector<std::pair<uint256, CDiskTxPos> > &vPosTxid,std::vector<std::pair<uint160, CExtDiskTxPos> > &vPosAddrid);
-
 // Internal stuff
 namespace {
     CBlockIndex *&pindexBestInvalid = g_chainstate.pindexBestInvalid;
@@ -285,7 +282,7 @@ CBlockIndex* FindForkInGlobalIndex(const CChain& chain, const CBlockLocator& loc
 
 std::unique_ptr<CCoinsViewDB> pcoinsdbview;
 std::unique_ptr<CCoinsViewCache> pcoinsTip;
-std::unique_ptr<CBlockTreeDB> pblocktree;
+std::unique_ptr<CAddrIndexBlockTreeDB> pblocktree;
 
 enum FlushStateMode {
     FLUSH_STATE_NONE,
@@ -1002,46 +999,6 @@ bool AcceptToMemoryPool(CTxMemPool& pool, CValidationState &state, const CTransa
     return AcceptToMemoryPoolWithTime(chainparams, pool, state, tx, pfMissingInputs, GetTime(), plTxnReplaced, bypass_limits, nAbsurdFee);
 }
 
-bool ReadTransaction(CTransactionRef& tx, const CDiskTxPos &pos, uint256 &hashBlock) {
-    CAutoFile file(OpenBlockFile(pos, true), SER_DISK, CLIENT_VERSION);
-    if (file.IsNull())
-        return error("%s: OpenBlockFile failed", __func__);
-    CBlockHeader header;
-    try {
-        file >> header;
-		fseek(file.Get(), pos.nTxOffset, SEEK_CUR);
-		file >> tx;
-    } catch (std::exception &e) {
-        LogPrintf("Upgrading txindex database... [%s]\n", e.what());
-        return error("%s() : deserialize or I/O error", __PRETTY_FUNCTION__);
-    }
-    hashBlock = header.GetHash();
-    return true;
-}
-
-bool FindTransactionsByDestination(const CTxDestination &dest, std::set<CExtDiskTxPos> &setpos) {
-    uint160 addrid;
-    const CKeyID *pkeyid = boost::get<CKeyID>(&dest);
-    if (pkeyid)
-        addrid = static_cast<uint160>(*pkeyid);
-    if (addrid.IsNull()) {
-        const CScriptID *pscriptid = boost::get<CScriptID>(&dest);
-        if (pscriptid)
-            addrid = static_cast<uint160>(*pscriptid);
-        }
-    if (addrid.IsNull())
-        return false;
-
-    LOCK(cs_main);
-    if (!fAddrIndex)
-        return false;
-    std::vector<CExtDiskTxPos> vPos;
-    if (!pblocktree->ReadAddrIndex(addrid, vPos))
-        return false;
-    setpos.insert(vPos.begin(), vPos.end());
-    return true;
-}
-
 /**
  * Return transaction in txOut, and if it was found inside a block, its hash is placed in hashBlock.
  * If blockIndex is provided, the transaction is fetched from the corresponding block.
@@ -1700,41 +1657,6 @@ static bool WriteUndoDataForBlock(const CBlockUndo& blockundo, CValidationState&
     return true;
 }
 
-// Index either: a) every data push >=8 bytes,  b) if no such pushes, the entire script
-void static BuildAddrIndex(const CScript &script, const CExtDiskTxPos &pos, std::vector<std::pair<uint160, CExtDiskTxPos> > &out)
-{
-    int outSize = out.size();
-    CScript::const_iterator pc = script.begin();
-    CScript::const_iterator pend = script.end();
-    std::vector<unsigned char> data;
-    opcodetype opcode;
-    bool fHaveData = false;
-    while (pc < pend) {
-        script.GetOp(pc, opcode, data);
-        if (0 <= opcode && opcode <= OP_PUSHDATA4 && data.size() >= 8) { // data element
-            uint160 addrid;
-            if (data.size() <= 20) {
-                memcpy(&addrid, &data[0], data.size());
-            } else {
-                addrid = Hash160(data);
-            }
-            //LogPrintf("BuildAddrIndex: add index ===== %s\n", addrid.GetHex());
-            out.push_back(std::make_pair(addrid, pos));
-            fHaveData = true;
-        }
-    }
-    if (!fHaveData) {
-        uint160 addrid = Hash160(script);
-        //LogPrintf("BuildAddrIndex: add index %s\n", addrid.GetHex());
-        out.push_back(std::make_pair(addrid, pos));
-    }
-	
-	if (outSize == out.size())
-	{
-		LogPrintf("BuildAddrIndex:get address from scriptPubkey failed. %s===========\n", ScriptToAsmStr(script));
-	}
-}
-
 static bool WriteTxIndexDataForBlock(CValidationState& state,std::vector<std::pair<uint256, CDiskTxPos> > &vPosTxid,std::vector<std::pair<uint160, CExtDiskTxPos> > &vPosAddrid)
 {
     if (!fTxIndex || !fAddrIndex) return true;
@@ -1744,19 +1666,6 @@ static bool WriteTxIndexDataForBlock(CValidationState& state,std::vector<std::pa
     }
 	if (!pblocktree->AddAddrIndex(vPosAddrid)){
 		return AbortNode(state, "Failed to write address index");
-	}
-    return true;
-}
-
-static bool EraseTxIndexDataForBlock(std::vector<std::pair<uint256, CDiskTxPos> > &vPosTxid,std::vector<std::pair<uint160, CExtDiskTxPos> > &vPosAddrid)
-{
-    if (!fTxIndex || !fAddrIndex) return true;
-
-    if (!pblocktree->EraseTxIndex(vPosTxid)) {
-        return false;
-    }
-	if (!pblocktree->EraseAddrIndex(vPosAddrid)){
-		return false;
 	}
     return true;
 }
